pyproject.toml
[project]
name = "capstone-image-recognition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.0"
dependencies = [
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
    "pandas>=1.3.0",
    "piexif>=1.1.3",
    "pillow>=9.0.0",
    "scipy>=1.7.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
